Adds get all products feature

+++ src/main/java/school/listecourses/exercicelistedecourses/application/product/commands/create/ProductCreateCommand.java
@@ -3,5 +3,5 @@ package school.listecourses.exercicelistedecourses.application.product.commands.
 public class ProductCreateCommand {
     public String name;
     public float price;
-    public int categoryId;
+    public long categoryId;
 }
