+++ src/main/java/school/listecourses/exercicelistedecourses/domain/Product.java
@@ -6,8 +6,8 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class Product {
-    private Long id;
+    private long id;
     private String name;
     private float price;
-    private int categoryId;
+    private long categoryId;
 }
